backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
mysql-connector-python==9.0.0
bcrypt==4.2.0
hypothesis==6.112.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
python-multipart==0.0.12
